improve and migrate tests for parseURLParams
- Add regression tests for major/major.minor version patterns
- Add tests for invalid input fallback behavior
- Migrate test file to TypeScript
- Export p5VersionStrings from parseUrlParams for test validation
- Use regex assertions to avoid breaking when p5Versions updates
